requirement.txt
blinker==1.3
chardet==2.3.0
click==6.6
colorama==0.3.7
cryptography==1.7.1
Flask==0.12.1
idna==2.2
itsdangerous==0.24
Jinja2==2.8
keyring==10.1
keyrings.alt==1.3
luma.core==1.12.0
luma.oled==3.3.0
MarkupSafe==0.23
Pillow==6.1.0
psutil==5.6.3
pyasn1==0.1.9
pycrypto==2.6.1
pyftdi==0.29.4
pygobject==3.22.0
pyinotify==0.9.6
pyOpenSSL==16.2.0
pyserial==3.4
python-apt==1.1.0b5
python-networkmanager==2.0.1
pyusb==1.0.2
pyxdg==0.25
requests==2.12.4
RPi.GPIO==0.7.0
rtmidi==2.3.4
python-rtmidi
numpy
SecretStorage==2.3.1
simplejson==3.10.0
six==1.12.0
smbus2==0.2.3
spidev==3.4
ssh-import-id==5.6
urllib3==1.19.1
Werkzeug==0.11.15
